# app/Http/Controllers/Api/AssetApiController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Jobs\GenerateAiTags;
use App\Models\Asset;
use App\Models\Setting;
use App\Models\Tag;
use App\Services\RekognitionService;
use App\Services\S3Service;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;

class AssetApiController extends Controller
{
    protected S3Service $s3Service;

    protected RekognitionService $rekognitionService;

    public function __construct(S3Service $s3Service, RekognitionService $rekognitionService)
    {
        $this->s3Service = $s3Service;
        $this->rekognitionService = $rekognitionService;
    }

    /**
     * Get paginated list of assets
     */
    public function index(Request $request)
    {
        $query = Asset::with(['tags', 'user']);

        // Apply filters
        if ($search = $request->input('search')) {
            $query->search($search);
        }

        if ($tagIds = $request->input('tags')) {
            $query->withTags(is_array($tagIds) ? $tagIds : explode(',', $tagIds));
        }

        if ($type = $request->input('type')) {
            $query->ofType($type);
        }

        if (Auth::user()->isAdmin() && $userId = $request->input('user')) {
            $query->byUser($userId);
        }

        if ($folder = $request->input('folder')) {
            $query->inFolder($folder);
        }

        // Apply sorting
        $sort = $request->input('sort', 'date_desc');
        switch ($sort) {
            case 'date_asc':
                $query->oldest('updated_at');
                break;
            case 'date_desc':
                $query->latest('updated_at');
                break;
            case 'upload_asc':
                $query->oldest('created_at');
                break;
            case 'upload_desc':
                $query->latest('created_at');
                break;
            case 'size_asc':
                $query->orderBy('size', 'asc');
                break;
            case 'size_desc':
                $query->orderBy('size', 'desc');
                break;
            case 'name_asc':
                $query->orderBy('filename', 'asc');
                break;
            case 'name_desc':
                $query->orderBy('filename', 'desc');
                break;
            case 's3key_asc':
                $query->orderBy('s3_key', 'asc');
                break;
            case 's3key_desc':
                $query->orderBy('s3_key', 'desc');
                break;
            default:
                $query->latest('updated_at');
        }

        $perPage = min($request->input('per_page', 24), 100);
        $assets = $query->paginate($perPage);

        return response()->json($assets);
    }

    /**
     * Upload new assets
     */
    public function store(Request $request)
    {
        $request->validate([
            'files.*' => 'required|file|max:512000', // 500MB max
        ]);

        $uploadedAssets = [];

        foreach ($request->file('files') as $file) {
            $fileData = $this->s3Service->uploadFile($file);

            $asset = Asset::create([
                's3_key' => $fileData['s3_key'],
                'filename' => $fileData['filename'],
                'mime_type' => $fileData['mime_type'],
                'size' => $fileData['size'],
                'width' => $fileData['width'],
                'height' => $fileData['height'],
                'user_id' => Auth::id(),
            ]);

            if ($asset->isImage()) {
                $thumbnailKey = $this->s3Service->generateThumbnail($asset->s3_key);
                if ($thumbnailKey) {
                    $asset->update(['thumbnail_s3_key' => $thumbnailKey]);
                }

                // Generate resized images
                try {
                    $resizedKeys = $this->s3Service->generateResizedImages($asset->s3_key);
                    if (! empty($resizedKeys)) {
                        $asset->update([
                            'resize_s_s3_key' => $resizedKeys['s'] ?? null,
                            'resize_m_s3_key' => $resizedKeys['m'] ?? null,
                            'resize_l_s3_key' => $resizedKeys['l'] ?? null,
                        ]);
                    }
                } catch (\Exception $e) {
                    \Log::error("Resize generation failed for {$asset->filename}: ".$e->getMessage());
                }

                if ($this->rekognitionService->isEnabled()) {
                    GenerateAiTags::dispatch($asset)->afterResponse();
                }
            }

            $uploadedAssets[] = $asset->fresh(['tags']);
        }

        return response()->json([
            'message' => count($uploadedAssets).' file(s) uploaded successfully',
            'data' => $uploadedAssets,
        ], 201);
    }

    /**
     * Get single asset details
     */
    public function show(Asset $asset)
    {
        $asset->load(['tags', 'user']);

        return response()->json($asset);
    }

    /**
     * Update asset metadata
     */
    public function update(Request $request, Asset $asset)
    {
        if (! Auth::user()->isAdmin() && $asset->user_id !== Auth::id()) {
            return response()->json(['message' => 'Unauthorized'], 403);
        }

        $request->validate([
            'alt_text' => 'nullable|string|max:500',
            'caption' => 'nullable|string|max:1000',
            'license_type' => 'nullable|string|max:255',
            'copyright' => 'nullable|string|max:500',
            'tags' => 'nullable|array',
            'tags.*' => 'string|max:50',
        ]);

        $asset->update(array_merge(
            $request->only(['alt_text', 'caption', 'license_type', 'copyright']),
            ['last_modified_by' => Auth::id()]
        ));

        // Handle tags - always sync, even if empty (to allow removing all tags)
        $tagIds = [];
        $tags = $request->input('tags', []);

        foreach ($tags as $tagName) {
            $tag = Tag::firstOrCreate(
                ['name' => strtolower(trim($tagName))],
                ['type' => 'user']
            );
            $tagIds[] = $tag->id;
        }

        $aiTagIds = $asset->aiTags()->pluck('tags.id')->toArray();
        $asset->tags()->sync(array_merge($aiTagIds, $tagIds));

        return response()->json([
            'message' => 'Asset updated successfully',
            'data' => $asset->fresh(['tags']),
        ]);
    }

    /**
     * Soft delete asset (move to trash)
     */
    public function destroy(Asset $asset)
    {
        if (! Auth::user()->isAdmin() && $asset->user_id !== Auth::id()) {
            return response()->json(['message' => 'Unauthorized'], 403);
        }

        $asset->delete();

        return response()->json([
            'message' => 'Asset moved to trash successfully',
        ]);
    }

    /**
     * Search assets (optimized for asset picker)
     */
    public function search(Request $request)
    {
        $query = Asset::with(['tags']);

        if ($search = $request->input('q')) {
            $query->search($search);
        }

        if ($tags = $request->input('tags')) {
            $tagIds = is_array($tags) ? $tags : explode(',', $tags);
            $query->withTags($tagIds);
        }

        if ($type = $request->input('type')) {
            $query->ofType($type);
        }

        if ($folder = $request->input('folder')) {
            $query->inFolder($folder);
        }

        // Apply sorting
        $sort = $request->input('sort', 'date_desc');
        switch ($sort) {
            case 'date_asc':
                $query->oldest('updated_at');
                break;
            case 'date_desc':
                $query->latest('updated_at');
                break;
            case 'upload_asc':
                $query->oldest('created_at');
                break;
            case 'upload_desc':
                $query->latest('created_at');
                break;
            case 'size_asc':
                $query->orderBy('size', 'asc');
                break;
            case 'size_desc':
                $query->orderBy('size', 'desc');
                break;
            case 'name_asc':
                $query->orderBy('filename', 'asc');
                break;
            case 'name_desc':
                $query->orderBy('filename', 'desc');
                break;
            case 's3key_asc':
                $query->orderBy('s3_key', 'asc');
                break;
            case 's3key_desc':
                $query->orderBy('s3_key', 'desc');
                break;
            default:
                $query->latest('updated_at');
        }

        $perPage = min($request->input('per_page', 24), 100);
        $assets = $query->paginate($perPage);

        return response()->json($assets);
    }

    /**
     * Get asset metadata by URL (public endpoint)
     */
    public function getMeta(Request $request)
    {
        // Check if meta endpoint is enabled
        if (! Setting::get('api_meta_endpoint_enabled', true)) {
            return response()->json(['message' => 'This endpoint is disabled.'], 403);
        }

        $request->validate([
            'url' => 'required|string|url',
        ]);

        $url = $request->input('url');
        $customBaseUrl = S3Service::getPublicBaseUrl().'/';
        $s3BaseUrl = config('filesystems.disks.s3.url').'/';

        // Extract S3 key by removing the base URL (supports both custom domain and S3 URLs)
        $s3Key = null;
        if (str_starts_with($url, $customBaseUrl)) {
            $s3Key = substr($url, strlen($customBaseUrl));
        } elseif ($s3BaseUrl !== $customBaseUrl && str_starts_with($url, $s3BaseUrl)) {
            $s3Key = substr($url, strlen($s3BaseUrl));
        }

        if ($s3Key === null) {
            return response()->json([
                'message' => 'URL does not match configured domain',
            ], 400);
        }

        // Find asset by s3_key
        $asset = Asset::where('s3_key', $s3Key)->first();

        if (! $asset) {
            return response()->json([
                'message' => 'Asset not found',
            ], 404);
        }

        // Return only metadata fields
        return response()->json([
            'alt_text' => $asset->alt_text,
            'caption' => $asset->caption,
            'license_type' => $asset->license_type,
            'copyright' => $asset->copyright,
            'filename' => $asset->filename,
            'url' => $asset->url,
        ]);
    }
}